Setup Settings View Controller
- Added About (Needs actual text)
- Displays the TerrorTorch logo (Needs actual logo)
- Added Link to www.goldenviking.org (Auto layout needs fixing)
- Names of all of the people that worked on this app (Needs all of the
names)

+++ TerrorTorch/Scenes/SoundBoxScene.swift
@@ -40,7 +40,7 @@ class SoundBoxScene: SKScene {
                     let url = NSURL.fileURLWithPath(path);
                     let audioAsset = AVURLAsset.URLAssetWithURL(url, options: nil);
                     let duration = audioAsset.duration;
-                    node.runAction(SKAction.rotateByAngle(6*M_PI, duration: CMTimeGetSeconds(duration)));
+                    node.runAction(SKAction.rotateByAngle(CGFloat(6*M_PI), duration: CMTimeGetSeconds(duration)));
                 }
             }
         }
